[Cassandra] [REST] moved stats request/response to model_rest.
